better connectivity detection
